3.2-crud/stocks_products/requirements.txt
django
Pillow
djangorestframework
psycopg2-binary
django-filter
pytest
pytest-django
model_bakery